+++ @kiva/kv-shop/src/index.ts
@@ -4,6 +4,7 @@ export * from './basketItems';
 export * from './basketTotals';
 export * from './basketVerification';
 export * from './checkoutStatus';
+export * from './managedAccount';
 export * from './oneTimeCheckout';
 export * from './receipt';
 export * from './shopError';